Align .NET EV Charger models with models.html specification
Agent-Logs-Url: https://github.com/Shane32/EnphaseAPI/sessions/f545e2d7-fbff-4826-8e75-6cf877d5b2b2

+++ Project/Models/EvCharger/EvChargerTelemetryInterval.cs
@@ -4,6 +4,6 @@ namespace Shane32.EnphaseAPI.Models;
 
 public class EvChargerTelemetryInterval
 {
-    [JsonPropertyName("consumption")] public long? Consumption { get; set; }
+    [JsonPropertyName("consumption")] public int? Consumption { get; set; }
     [JsonPropertyName("end_at")] public long? EndAt { get; set; }
 }
